server: create separate class for handling conns
create log function for thread-safe printing
