+++ CMakeLists.txt
@@ -33,7 +33,6 @@ add_executable(hikup-server src/server/main.cpp
         src/server/ClientInfo.hpp
         src/server/ConnectionServer.cpp
         src/server/ConnectionServer.hpp
-        src/server/utilServer.cpp
         src/server/HTTPFileServer.cpp
         src/server/HTTPFileServer.hpp
         src/server/includes/mongoose.cpp
@@ -43,7 +42,13 @@ add_executable(hikup-server src/server/main.cpp
         src/shared/FileInfo.hpp
         src/shared/utils.cpp
         src/shared/utils.cpp
-        src/shared/utils.hpp)
+        src/shared/utils.hpp
+        src/server/ConnectionHandler.cpp
+        src/server/ConnectionHandler.hpp
+        src/server/Settings.hpp
+        src/server/Settings.cpp
+        src/server/utils.cpp
+        src/server/utils.hpp)
 
 target_include_directories(hikup PRIVATE ${LIBSODIUM_INCLUDE_DIRS})
 target_link_libraries(hikup ${LIBSODIUM_LIBRARIES})
