+++ src/server/ConnectionServer.hpp
@@ -39,10 +39,7 @@ public:
 
 	void resizeBuffer ( unsigned long newSize );
 
-	struct ConnectionSpeed {
-		size_t uploadSpeed = 0;
-		size_t downloadSpeed = 0;
-	};
+	[[nodiscard]] bool isActive () const;
 
 private:
 	struct KeyPair {
@@ -54,7 +51,6 @@ private:
 	KeyPair _keyPair;
 	ClientInfo _clientInfo;
 	std::vector<std::string> _messagesBuffer;
-	ConnectionSpeed _connectionSpeed;
 
 
 	long int _sizeOfPreviousMessage = 0;
@@ -74,5 +70,4 @@ private:
 
 	void secretSeal ( std::string& message ) const;
 
-	void _handleConnSpeedTest ();
 };
