+++ src/server/utils.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+
+namespace Utils {
+    void log ( const std::string& message, bool newline = true );
+    void elog ( const std::string& message, bool newline = true );
+}
